Moved the intro to its own file
