+++ src/viewer/MaterialBitmapEngravesViewer.tsx
@@ -6,7 +6,7 @@ import {
 } from '../lib/glowforgeUnits';
 import { PluginMaterial } from '../material/materialPlugin';
 import { precisionRound } from '../lib/utils';
-import './Viewer.css';
+import './MaterialViewer.css';
 
 export type MaterialBitmapEngravesViewerProps = {
   bitmaps: PluginMaterial['bitmaps'];
